+++ sdk/src/main/java/net/siegmar/billomat4j/sdk/domain/settings/Settings.java
@@ -102,6 +102,10 @@ public class Settings {
     private Integer dueDays;
     private Integer offerValidityDays;
     private BccAddresses bccAddresses;
+    private String bgcolor;
+    private String color1;
+    private String color2;
+    private String color3;
 
     public String getInvoiceIntro() {
         return invoiceIntro;
@@ -473,6 +477,38 @@ public class Settings {
 
     public void setBccAddresses(final BccAddresses bccAddresses) {
         this.bccAddresses = bccAddresses;
+    }
+
+    public String getBgcolor() {
+        return bgcolor;
+    }
+
+    public void setBgcolor(final String bgcolor) {
+        this.bgcolor = bgcolor;
+    }
+
+    public String getColor1() {
+        return color1;
+    }
+
+    public void setColor1(final String color1) {
+        this.color1 = color1;
+    }
+
+    public String getColor2() {
+        return color2;
+    }
+
+    public void setColor2(final String color2) {
+        this.color2 = color2;
+    }
+
+    public String getColor3() {
+        return color3;
+    }
+
+    public void setColor3(final String color3) {
+        this.color3 = color3;
     }
 
     @Override