+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Removed setters for read-only elements in settings
 * Added support for archiving customers
 * Added support for paid_amount and open_amount in invoices
+* Added support for bgcolor, color1, color2, color3 in settings
 
 ## 1.0.3 (2013-02-09)